Fix issue with health bar
Fixes an issue where the health bar wouldn't deplete for one of
the enemies. Because we were using GameObject.Find to get the
reference to the health bar, it would return the same health bar
for each enemy. Once the enemy was destroyed, it would not be able
to find it and would throw an error. This uses transform.Find,
which looks for an object that is local to the current gameObject