Get free directions in Floor
Uses the Floor grid to get free directions instead of the Room.
The floor has a better idea of what edges are free because it knows
what tiles are adjacent to a room even if they aren't connected.
This will give us less buggy mazes and better separation of concerns